Implement covariance pooling and associated SIGReg loss in JEPA model
- Refactored the export_massspec_embeddings.py script to replace encoder_embedding with covariance_embedding.
- Updated the _encode_splits function to compute covariance embeddings using the new covariance_pooler.
- Added tests in test_pretrain.py to verify the contribution of covariance pooling to the JEPA loss and ensure it is not slotwise.
- Modified train.py to include loading and saving of covariance_sigreg parameters.
- Enhanced training.py to support configuration options for covariance pooling and its regularization.

+++ configs/gems_small_norm.py
@@ -92,7 +92,7 @@ def get_config() -> config_dict.ConfigDict:
     cfg.jepa_target_normalization = "zscore"
     cfg.jepa_target_layers = [3, 5, 8, 12]
     cfg.grad_clip_norm = 1.0
-    cfg.autocast_dtype = "fp32"
+    cfg.autocast_dtype = "bf16"
     cfg.compile_mode = "reduce-overhead"
     cfg.representation_regularizer = "slot-sigreg-enc"
     cfg.train_covariance_pooling = True
